Integrated test code file
